requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.89.1
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
pydantic==1.10.4
pyee==9.0.4
python-dotenv==0.21.0
slack-bolt==1.16.1
slack-sdk==3.19.5
slackclient==2.9.4
slackeventsapi==3.0.1
sniffio==1.3.0
starlette==0.22.0
tinydb==4.7.0
typing_extensions==4.4.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
